feat(drag-bar): added drag bar to adjust width of code editor and output console (#4)
Added a drag bar at the centre of the page in between the code editor and output console. The drag bar can be clicked and dragged to adjust the width of the code editor and output console. The widths of each is limited to a minimum of 30% of the page width and maximum of 64.2% of page width. The icon of the mouse/pointer changes when hovering over the drag bar to signal to the user that they can start dragging the component.