feat(interpreter-parser): added basic parser and syntax error functionality (#7)
Created a basic version of the parser (based on a recursive-descent parser) which at the moment can parse through arithmetic expressions including the +, -, * and / operations and supports bracketed precedence as well as operation precedence. The parser generates an appropriate Abstract Syntax Tree made up of ParseResults which group information about the nodes in the tree and the errors (if they exist). Also added a syntax error detector in the parser as well as a new PositionTracker class to track the index, line and column number of the tokens for better feedback in error messages. Unit tests were written for all these new features related to the parser covering the different supported expressions and erroneous expressions.