feat(parser-logical-operations): added logical operation support in parser (#10)
The parser now supports KEYWORDS which in this change includes: TRUE, FALSE, AND, OR and NOT. These allow these boolean operations to be supported correctly by the parser to distinguish between IDENTIFIERs which are regular variables. The parser also supports creating ASTs for comparison expressions using the following operators: ==, >, <, >=, <=. There was also refactoring done to reorganise the constants files in their own constants folder.