feat(output-function): added support for built-in functions output and toString (#18)
The language now supports two new built-in functions: output and toString. The output function takes in a StringType as an argument and outputs that string to both the console and shell output. The toString function takes in either a: NumberType, StringType or ListType and converts them to a StringType and returns that value. Some examples shown below:

- toString("hello world!"): Console Output: and Shell Output: hello world!
- toString(1): Console Output: and Shell Output: 1
- toString([1, 2, 3]): Console Output: and Shell Output: [1, 2, 3]
- output("hello world!"): Console Output: hello world! and Shell Output: hello world!
- output(toString([1, 2, 3])): Console Output: [1, 2, 3] and Shell Output: [1, 2, 3]