feat(incorporate-interpreter): added interpreter to play button mechanism (#21)
The play button is now able to read the Code window source code and run it through the Pseudo interpreter to generate the appropriate console output in the Output window. Note: at the moment, empty spaces that are added to indent code, seem to through invalid character errors as a result of the representation of the innerText in the DOM. This still needs to be fixed, but other than that, the IDE understands the correct syntax.